refactor: 숫자 생성을 LottoMachine이 아닌 NumbersGenerator에게 위임 - 로또의 생성은 LottoMachine에서 하되, 중복되지 않은 6개의 숫자를 생성하는 역할은 NumbersGenerator 객체를 생성하여 위임하였다.
